+++ bestmobabot/enums.py
@@ -4,8 +4,3 @@ from enum import Enum
 class BattleType(Enum):
     ARENA = 'arena'
     GRAND = 'grand'
-
-
-__all__ = [
-    'BattleType',
-]