+++ bestmobabot/vk.py
@@ -4,11 +4,11 @@ VK.com API wrapper.
 
 import contextlib
 import re
-from typing import Any, Dict, Iterable
+from typing import Iterable
 
 import requests
+from loguru import logger
 
-from bestmobabot.logging_ import logger
 from bestmobabot.settings import Settings
 
 
@@ -19,7 +19,12 @@ class VK(contextlib.AbstractContextManager):
 
     def __init__(self, settings: Settings):
         self.session = requests.Session()
-        self.params = {'access_token': settings.vk.access_token, 'owner_id': '-116039030', 'count': '5', 'v': '5.69'}
+        self.params = {
+            'access_token': settings.vk.access_token,
+            'owner_id': '-116039030',
+            'count': '5',
+            'v': '5.69',
+        }
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.session.__exit__(exc_type, exc_val, exc_tb)
@@ -31,12 +36,12 @@ class VK(contextlib.AbstractContextManager):
             response.raise_for_status()
             payload = response.json()
 
-        for item in payload['response']['items']:  # type: Dict[str, Any]
+        for item in payload['response']['items']:
             yield from self.GIFT_RE.findall(item['text'])
             for url in self.VK_CC_RE.findall(item['text']):
                 # HEAD is not supported by VK.com.
                 with self.session.get(url, stream=True) as response:
                     yield from self.GIFT_RE.findall(response.url)
-            for attachment in item['attachments']:  # type: Dict[str, Any]
+            for attachment in item['attachments']:
                 if attachment['type'] == 'link':
                     yield from self.GIFT_RE.findall(attachment['link']['url'])